+++ lib/src/models/change.dart
@@ -1,11 +1,13 @@
 import '../lib/coinbase_enums.dart';
 import '../lib/custom_json_converters.dart';
 import 'package:json_annotation/json_annotation.dart';
+import '../lib/websocket_response.dart';
+
 
 part 'change.g.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake, explicitToJson: true)
-class Change {
+class Change extends WebsocketResponse {
   final String? type;
   final DateTime? time;
   final int? sequence;