+++ lib/src/models/heartbeat.dart
@@ -1,9 +1,10 @@
+import '../lib/websocket_response.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'heartbeat.g.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake)
-class Heartbeat {
+class Heartbeat extends WebsocketResponse {
   final String? type;
   final int? sequence;
   final int? lastTradeId;