+++ lib/src/models/open.dart
@@ -1,11 +1,12 @@
 import '../lib/coinbase_enums.dart';
 import '../lib/custom_json_converters.dart';
+import '../lib/websocket_response.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'open.g.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake, explicitToJson: true)
-class Open {
+class Open extends WebsocketResponse {
   final String? type;
   final DateTime? time;
   final String? productId;