+++ lib/src/models/status.dart
@@ -1,3 +1,4 @@
+import '../lib/websocket_response.dart';
 import '../models/currency.dart';
 import '../models/product.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -5,7 +6,7 @@ import 'package:json_annotation/json_annotation.dart';
 part 'status.g.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake, explicitToJson: true)
-class Status {
+class Status extends WebsocketResponse {
   final String? type;
   final List<Product?>? products;
   final List<Currency?>? currencies;