+++ lib/src/models/websocket_error.dart
@@ -1,0 +1,7 @@
+import '../lib/websocket_response.dart';
+
+class WebsocketError extends WebsocketResponse {
+  final String? message;
+
+  WebsocketError({this.message});
+}