+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+- Add an example flutter application
+- Make websocket responses subclasses of a WebsocketResponse class
+
 ## 0.0.2
 - Run dart format